Improved Data Accuracy and GPS sample
-Changed SD to append every 200 ms instead of every 100 ms
-Changed GPS sample rate to every 1 sec instead of 10 seconds
-Instead of relying on delays to keep track of time, use hardware timer to keep track of exact time and not lose sync

+++ Code/demo/freeRTOS/Definitions.h
@@ -5,7 +5,7 @@ extern bool dataActive;
 
 extern float latitude, longitude;
 extern unsigned long gps_date, gps_time, age;
-extern float dataTime;
+extern double dataTime;
 //Task Handle, used to suspend start tasks using others
 //extern static TaskHandle_t taskDATA;
